requirements.txt
# ==============================================================================
# 🐾 Evcil Hayvan Platformu - Production Requirements
# ==============================================================================
# Bu dosya, production ortamında çalışacak temel paketleri içerir.
# Her paket, platform güvenliği ve performansı için dikkatli seçilmiştir.
# ==============================================================================

# ==============================================================================
# 🎯 CORE DJANGO FRAMEWORK - Projenin kalbi
# ==============================================================================
Django==4.2.16                     # Django LTS version (django-celery-beat uyumlu)
djangorestframework==3.14.0        # RESTful API development harmony'si
django-environ==0.11.2             # Environment variables güvenli yönetimi
python-decouple==3.8               # Environment configuration management
setuptools>=65.0.0                 # Python package tools (pkg_resources için)

# ==============================================================================
# 🔗 DATABASE & CACHE - Veri katmanı sevgisi
# ==============================================================================
psycopg2-binary==2.9.10           # PostgreSQL adapter (hayvan hikayelerinin evi)
redis==5.0.1                       # In-memory cache (hızlı sevgi dağıtımı)
django-redis==5.4.0                # Django-Redis entegrasyonu

# ==============================================================================
# 🔐 GÜVENLİK & AUTH - Kalp koruması sistemi
# ==============================================================================
djangorestframework-simplejwt==5.3.0   # JWT authentication (güvenli token'lar)
django-cors-headers==4.7.0             # CORS policy management (frontend-backend dostluğu)

# Security packages
django-csp==3.7                    # Content Security Policy
django-ratelimit==4.1.0                # Rate limiting (DDoS koruması)
django-ipware==5.0.0               # IP adresi güvenliği
cryptography==41.0.7                # Şifreleme ve güvenlik

# ==============================================================================
# 📁 DOSYA & MEDYA YÖNETİMİ - Sevgi fotoğrafları için
# ==============================================================================
Pillow==11.2.1                     # Image processing (hayvan fotoğrafları güzelliği)
django-storages==1.14.4            # Cloud storage backends (AWS S3 desteği)
boto3==1.35.83                     # AWS SDK (cloud dosya yönetimi)

# ==============================================================================
# 🔍 FİLTRELEME & ARAMA - Akıllı eşleştirme
# ==============================================================================
django-filter==23.4                # Advanced filtering (hayvan arama sistemi)
django-elasticsearch-dsl==8.0      # Elasticsearch integration (güçlü arama)
elasticsearch==8.11.1              # Search engine (hayvan profil araması)

# ==============================================================================
# 🔄 ASYNC TASK MANAGEMENT - Arka plan sevgi işçileri
# ==============================================================================
celery==5.5.0                      # Distributed task queue (async görevler)
django-celery-beat==2.6.0          # Periodic task scheduler (Django 5.2 uyumlu versiyon)
kombu>=5.5.2,<5.6                  # Message transport (celery uyumlu versiyon)

# ==============================================================================
# 🌐 API & SERIALIZATION - Veri harmonisi
# ==============================================================================
djangorestframework-gis==1.0       # Geographic REST framework (konum servisleri)
drf-spectacular==0.27.0             # OpenAPI 3.0 schema generation (API docs)
django-crispy-forms==2.1           # Beautiful form rendering

# ==============================================================================
# 📊 MONİTORİNG & LOGGİNG - Platform sağlık nabzı
# ==============================================================================
sentry-sdk==1.40.6                 # Error tracking (hata izleme sistemi)
django-health-check==3.17.0        # Health check endpoints (sistem durumu)

# ==============================================================================
# 🌍 INTERNATIONALIZATION - Küresel sevgi
# ==============================================================================
django-modeltranslation==0.18.11   # Model field translation (çoklu dil)

# ==============================================================================
# 📧 İLETİŞİM & BİLDİRİMLER - Sevgi mesajları
# ==============================================================================
django-anymail==10.2               # Email service provider integration
requests==2.31.0                   # HTTP library (external API calls)

# ==============================================================================
# 🛡️ GÜVENLİK ENHANCEMENTs - Ekstra koruma
# ==============================================================================
django-secure==1.0.2               # Security middleware collection

# ==============================================================================
# ⚡ PERFORMANCE - Hız optimizasyonu
# ==============================================================================
django-cachalot==2.6.1             # ORM cache (database query optimization)
whitenoise==6.6.0                  # Static file serving (production ready)

# ==============================================================================
# 🔧 UTILITIES - Geliştirme yardımcıları
# ==============================================================================
python-dateutil==2.8.2             # Date/time utilities
pytz==2023.3                       # Timezone handling (global time support)
Faker==20.1.0                      # Test data generation (sample pet data)

# ==============================================================================
# 🌊 WEBSOCKET & ASYNC - Gerçek zamanlı iletişim
# ==============================================================================
channels==4.0.0                    # WebSocket support (mesajlaşma için)
channels-redis==4.1.0              # Redis channel layer (WebSocket backend)

# ==============================================================================
# 🔧 ADDITIONAL DEPENDENCIES - Ek bağımlılıklar
# ==============================================================================
dj-database-url==2.1.0             # Database URL parsing (environment flexibility)
gunicorn==21.2.0                   # WSGI server (production deployment)
